backend/chat_service/requirements.txt
Flask
Flask-SocketIO
eventlet # or gevent, for async mode with Flask-SocketIO
redis # For Flask-SocketIO message queue
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
PyJWT
Flask-JWT-Extended 